refactor cli app refactor core threads write workflow for common video
~~拉了托大的（x~~

+++ bilicore/core.py
@@ -12,6 +12,63 @@ from bilicore.parser import select_quality
 from bilicore.utils import filename_escape, merge_avfile, call_ffmpeg
 
 
+class ThreadProgressMixin:
+    def __init__(self) -> None:
+        self.__curr = 0
+        self.__total = 0
+        self.__progress_text = ""
+        self.__report_lock = threading.Lock()
+        self.__exceptions: list[Exception] = []
+        self._report_progress(0, 0, "pending")
+
+    def _report_progress(
+        self,
+        curr: Optional[int] = None,
+        total: Optional[int] = None,
+        pgr_text: Optional[str] = None,
+    ):
+        with self.__report_lock:
+            if curr is not None:
+                self.__curr = curr
+            if total is not None:
+                self.__total = total
+            if pgr_text is not None:
+                self.__progress_text = pgr_text
+
+    def _report_exception(self, exc: Exception):
+        with self.__report_lock:
+            self.__exceptions.append(exc)
+
+    def observe(self):
+        with self.__report_lock:
+            return self.__curr, self.__total, self.__progress_text
+
+    @property
+    def exceptions(self):
+        with self.__report_lock:
+            return self.__exceptions.copy()
+
+    def _progress_hook(
+        self,
+        curr: Optional[int] = None,
+        total: Optional[int] = None,
+        offset: int = 0,
+    ):
+        if curr:
+            self._report_progress(curr=curr + offset)
+        if total:
+            self._report_progress(total=total + offset)
+
+    def _run_wrapped(self, worker: Callable[[], Any]):
+        # 这个写法有点傻逼的
+        try:
+            self._report_progress(pgr_text="starting")
+            worker()
+        except Exception as e:
+            self._report_exception(e)
+            self._report_progress(pgr_text="errored")
+
+
 class ThreadUtilsMixin:
     @staticmethod
     def _dstream(
@@ -40,19 +97,22 @@ class ThreadUtilsMixin:
         with open(file, "wb+") as fp:
             fp.write(data)
 
-    @staticmethod
-    def _run(self):  # pylint: disable=W0211
-        # 这个写法有点傻逼的
-        try:
-            self._progress_text = "starting"
-            self._worker()
-            self._progress_text = "done"
-        except Exception as e:
-            self.exception = e
-            self._progress_text = "errored"
 
+class SingleVideoThread(threading.Thread, ThreadUtilsMixin, ThreadProgressMixin):
+    VALID_OPTIONS = (
+        # 下载选项
+        "audio_only",
+        "video_quality",
+        "audio_quality",
+        "video_codec",
+        "subtitle_lang",
+        "subtitle_format",
+        # 预处理数据
+        "stream_data",
+        "video_data",
+        "player_info",
+    )
 
-class SingleVideoThread(threading.Thread, ThreadUtilsMixin):
     def __init__(
         self,
         apis: APIContainer,
@@ -64,28 +124,31 @@ class SingleVideoThread(threading.Thread, ThreadUtilsMixin):
         **options,
     ) -> None:
         super().__init__(daemon=True)
+        ThreadProgressMixin.__init__(self)
         if (avid is None) == (bvid is None):
             raise ValueError("need ONE between avid and bvid")
         self._cid = cid
         self._id = remove_none({"avid": avid, "bvid": bvid})
         self._apis = apis
         self._savedir = savedir
+
+        options = {k: v for k, v in options.items() if k in self.VALID_OPTIONS}
         self._audio_only = bool(options.get("audio_only", False))
-        self._video_data: Optional[dict[str, Any]] = options.get("video_data")
-        self._streams: Optional[dict[str, Any]] = options.get("stream_data")
-        self._player_info: Optional[dict[str, Any]] = options.get("player_info")
         self._vq: str | int = options.get("video_quality", "max")
         self._vc: Literal["avc", "hevc"] = options.get("video_codec", "avc")
         self._aq: str | int = options.get("audio_quality", "max")
         self._sv: Optional[str | Literal["all"]] = options.get("subtitle_lang")
         self._sf: Literal["vtt", "srt", "lrc"] = options.get("subtitle_format", "vtt")
-        self._progress_text = "pending"
-        self.exception: Optional[Exception] = None
+
+        self._video_data: Optional[dict[str, Any]] = options.get("video_data")
+        self._streams: Optional[dict[str, Any]] = options.get("stream_data")
+        self._player_info: Optional[dict[str, Any]] = options.get("player_info")
 
     def run(self):
-        self._run(self)
+        self._run_wrapped(self._worker)
 
     def _worker(self):
+        self._report_progress(pgr_text="collecting data")
         if self._video_data is None:
             self._video_data = self._apis.video.get_video_detail(**self._id)
         vdata = self._video_data
@@ -108,15 +171,26 @@ class SingleVideoThread(threading.Thread, ThreadUtilsMixin):
         vstream, astream = select_quality(
             streams, aq=self._aq, vq=self._vq, enc=self._vc
         )
+        # 处理没有音轨的情况
+        no_audio = astream is None
+        if no_audio and self._audio_only:
+            self._report_progress(pgr_text="terminated: no audio stream to download")
+            return
         title = vdata["title"]
         ptitle = vdata["pages"][pindex]["part"]
         vqid = vstream["id"]
-        aqid = astream["id"]
+        aqid = -1 if no_audio else astream["id"]
         is_lossless = aqid == 30251
+        # 生成文件名
         vtmpfile = os.path.join(self._savedir, f"{bvid}_{cid}_{vqid}_videostream.m4v")
-        atmpfile = os.path.join(
-            self._savedir,
-            f"{bvid}_{cid}_{aqid}_videostream" + (".flac" if is_lossless else ".m4a"),
+        atmpfile = (
+            ""
+            if no_audio
+            else os.path.join(
+                self._savedir,
+                f"{bvid}_{cid}_{aqid}_videostream"
+                + (".flac" if is_lossless else ".m4a"),
+            )
         )
         finalfile = os.path.join(
             self._savedir,
@@ -140,11 +214,12 @@ class SingleVideoThread(threading.Thread, ThreadUtilsMixin):
             ),
         )
         if os.path.isfile(finalfile):
-            self._progress_text = "skipped"
+            self._report_progress("skipped: final file already exists")
             return
         # 字幕
         subfile = os.path.join(self._savedir, finalfile + ".{lan}" + f".{self._sf}")
         if self._sv is not None:
+            self._report_progress(pgr_text="fetching subtitle")
             for sub in subtitles:
                 if self._sv == "all" or self._sv.lower() == sub["lan"].lower():
                     self._dsubt(
@@ -153,13 +228,18 @@ class SingleVideoThread(threading.Thread, ThreadUtilsMixin):
                         self._sf,
                     )
         vurls = [vstream["base_url"]] + vstream["backup_url"]
-        aurls = [astream["base_url"]] + astream["backup_url"]
-        ahook = functools.partial(self._progress_hook, name="audio")
-        self._dstream(aurls, atmpfile, ahook, apis=self._apis)
+        self._report_progress(pgr_text="fetching stream")
+        if not no_audio:
+            aurls = [astream["base_url"]] + astream["backup_url"]
+            ahook = self._progress_hook
+
+            self._dstream(aurls, atmpfile, ahook, apis=self._apis)
+        # 仅音轨的分岔
         if self._audio_only:
             if is_lossless:
                 os.rename(atmpfile, finalfile)
             else:
+                self._report_progress(pgr_text="converting")
                 call_ffmpeg(
                     "-i",
                     atmpfile,
@@ -168,12 +248,22 @@ class SingleVideoThread(threading.Thread, ThreadUtilsMixin):
                     finalfile,
                 )
                 os.remove(atmpfile)
+            self._report_progress(pgr_text="done")
             return
-        vhook = functools.partial(self._progress_hook, name="video")
+        # 普通视频的分岔
+        vhook = functools.partial(
+            self._progress_hook,
+            offset=(os.path.getsize(atmpfile) if os.path.isfile(atmpfile) else 0),
+        )
         self._dstream(vurls, vtmpfile, vhook, apis=self._apis)
-        merge_avfile(atmpfile, vtmpfile, finalfile)
-        os.remove(atmpfile)
+        self._report_progress(pgr_text="merging")
+        if no_audio:
+            call_ffmpeg("-i", vtmpfile, "-vcodec", "copy", finalfile)
+        else:
+            merge_avfile(atmpfile, vtmpfile, finalfile)
+            os.remove(atmpfile)
         os.remove(vtmpfile)
+        self._report_progress(pgr_text="done")
 
     def _dsubt(self, url, file, fmt: Literal["vtt", "srt", "lrc"]):
         subt = self._apis.session.get(url, headers=self._apis.DEFAULT_HEADERS).json()
@@ -186,31 +276,29 @@ class SingleVideoThread(threading.Thread, ThreadUtilsMixin):
                 case "vtt":
                     f.write(subtitle.bcc2vtt(subt))
 
-    def _progress_hook(self, curr: Optional[int], total: Optional[int], name: str):
-        self._progress_text = (
-            f"{name}: --%"
-            if curr is None or total is None
-            else f"{name}: {curr/total*100:.2f}%"
-        )
 
-    def observe(self):
-        return self._progress_text
+class SingleAudioThread(threading.Thread, ThreadUtilsMixin, ThreadProgressMixin):
+    VALID_OPTIONS = (
+        "quality",
+        "lyrics",
+        "cover",
+        "audio_data",
+    )
 
-
-class SingleAudioThread(threading.Thread, ThreadUtilsMixin):
     def __init__(self, apis: APIContainer, auid, savedir, **options) -> None:
         super().__init__(daemon=True)
+        ThreadProgressMixin.__init__(self)
         self._apis = apis
         self._auid = auid
         self._savedir = savedir
+        options = {k: v for k, v in options.items() if k in self.VALID_OPTIONS}
         self._quality: Optional[Literal[0, 1, 2, 3]] = options.get("quality", 3)
+        self._need_lrc = bool(options.get("lyrics", False))
+        self._need_cover = bool(options.get("cover", False))
         self._info: Optional[dict[str, Any]] = options.get("audio_data")
-        self._progress_text = "pending"
-        self._need_lrc = bool(options.get("need_lrc", False))
-        self._need_cover = bool(options.get("need_cover", False))
-        self.exception: Optional[Exception] = None
 
     def _worker(self):
+        self._report_progress(pgr_text="collecting data")
         info = self._info if self._info else self._apis.audio.get_info(self._auid)
         auid = info["id"]
         stream = self._apis.audio.get_stream(auid=auid, quality=self._quality)
@@ -233,12 +321,14 @@ class SingleAudioThread(threading.Thread, ThreadUtilsMixin):
         if (cover := info.get("cover")) and self._need_cover:
             self._dfile(cover, finalfile + os.path.splitext(cover)[1], self._apis)
         if os.path.isfile(finalfile):
-            self._progress_text = "skipped"
+            self._report_progress("skipped")
             return
+        
+        self._report_progress(pgr_text="fetching stream")
         self._dstream(
             stream["cdns"],
             tmpfile,
-            functools.partial(self._progress_hook, name="stream"),
+            self._progress_hook,
             apis=self._apis,
         )
         if is_lossless:
@@ -252,33 +342,30 @@ class SingleAudioThread(threading.Thread, ThreadUtilsMixin):
                 finalfile,
             )
             os.remove(tmpfile)
+        self._report_progress(pgr_text="done")
 
     def run(self):
-        self._run(self)
-
-    def _progress_hook(self, curr: Optional[int], total: Optional[int], name: str):
-        self._progress_text = (
-            f"{name}: --%"
-            if curr is None or total is None
-            else f"{name}: {curr/total*100:.2f}%"
-        )
-
-    def observe(self):
-        return self._progress_text
+        self._run_wrapped(self._worker)
 
 
-class SingleMangaChapterThread(threading.Thread, ThreadUtilsMixin):
+class SingleMangaChapterThread(threading.Thread, ThreadUtilsMixin, ThreadProgressMixin):
+    VALID_OPTIONS = (
+        "create_childfolder",
+        "ep_data",
+    )
+
     def __init__(self, apis: APIContainer, epid: int, savedir: str, **options) -> None:
         super().__init__(daemon=True)
+        ThreadProgressMixin.__init__(self)
         self._apis = apis
         self._epid = epid
         self._savedir = savedir
-        self._ep_data: Optional[dict[str, Any]] = options.get("ep_data")
+        options = {k: v for k, v in options.items() if k in self.VALID_OPTIONS}
         self._create_childfolder = bool(options.get("create_childfolder", True))
-        self._progress_text = "pending"
-        self.exception: Optional[Exception] = None
+        self._ep_data: Optional[dict[str, Any]] = options.get("ep_data")
 
     def _worker(self):
+        self._report_progress(pgr_text="collecting data")
         epinfo = (
             self._ep_data
             if self._ep_data
@@ -295,7 +382,7 @@ class SingleMangaChapterThread(threading.Thread, ThreadUtilsMixin):
         paths = [i["path"] for i in index["images"]]
         tokens = self._apis.manga.get_image_token(*paths)
         urls = [i["url"] + "?token=" + i["token"] for i in tokens]
-        self._progress_text = f"0 / {len(urls)}"
+        self._report_progress(0, len(urls), pgr_text="downloading")
         with ThreadPoolExecutor(max_workers=8) as executor:
             futures = [
                 executor.submit(
@@ -309,9 +396,10 @@ class SingleMangaChapterThread(threading.Thread, ThreadUtilsMixin):
             for i, future in enumerate(as_completed(futures)):
                 try:
                     future.result()
-                    self._progress_text = f"{i+1} / {len(urls)}"
+                    self._report_progress(i + 1, len(urls))
                 except Exception as e:
-                    self.exception = e
+                    self._report_exception(e)
+        self._report_progress(pgr_text="done")
 
     def run(self):
-        self._run(self)
+        self._run_wrapped(self._worker)
